+++ src/H2DE_abs_rect.cpp
@@ -1,4 +1,4 @@
-#include "H2DE_utils.h"
+#include "H2DE/H2DE_utils.h"
 
 H2DE_AbsRect::operator SDL_Rect() const {
     return { x, y, w, h };