+++ src/H2DE_abs_size.cpp
@@ -1,4 +1,4 @@
-#include "H2DE_utils.h"
+#include "H2DE/H2DE_utils.h"
 
 // METHODS
 H2DE_AbsRect H2DE_AbsSize::makeRect(H2DE_AbsPos pos) const {