+++ src/H2DE_engine.cpp
@@ -1,4 +1,4 @@
-#include "H2DE_engine.h"
+#include "H2DE/H2DE_engine.h"
 
 // INIT
 H2DE_Engine::H2DE_Engine(H2DE_EngineData d) : data(d), fps(data.fps) {
@@ -99,8 +99,10 @@ void H2DE_Engine::click(int x, int y) {
         H2DE_ButtonData* btnData = H2DE_GetButtonData(button);
         if (!btnData->onClick) continue;
 
+        H2DE_TextureData btnTextureData = btnData->texture->getData();
+
         H2DE_AbsPos pos = H2DE_AbsPos{ static_cast<int>(btnData->pos.x * blockSize), static_cast<int>(btnData->pos.y * blockSize) };
-        H2DE_AbsSize size = H2DE_AbsSize{ static_cast<int>(btnData->texture.size.w * blockSize), static_cast<int>(btnData->texture.size.h * blockSize) };
+        H2DE_AbsSize size = H2DE_AbsSize{ static_cast<int>(btnTextureData.size.w * blockSize), static_cast<int>(btnTextureData.size.h * blockSize) };
         H2DE_AbsRect btnRect = pos.makeRect(size);
 
         if (btnRect.contains(clickPos)) btnData->onClick();