+++ src/H2DE_functions.cpp
@@ -1,4 +1,4 @@
-#include "H2DE_utils.h"
+#include "H2DE/H2DE_utils.h"
 
 float H2DE_Lerp(float min, float max, float blend) {
     return min + blend * (max - min);