+++ src/H2DE_level_hitbox.cpp
@@ -1,4 +1,4 @@
-#include "H2DE_utils.h"
+#include "H2DE/H2DE_utils.h"
 
 // LEVEL POS OPERATIONS
 H2DE_LevelRect H2DE_LevelRect::operator+(const H2DE_LevelPos& pos) const {