+++ src/H2DE_level_pos.cpp
@@ -1,4 +1,4 @@
-#include "H2DE_utils.h"
+#include "H2DE/H2DE_utils.h"
 
 // LEVEL POS OPERATIONS
 H2DE_LevelPos H2DE_LevelPos::operator+(const H2DE_LevelPos& other) const {