+++ src/H2DE_ostream.cpp
@@ -1,4 +1,4 @@
-#include "H2DE_utils.h"
+#include "H2DE/H2DE_utils.h"
 
 std::ostream& operator<<(std::ostream& os, const H2DE_AbsPos& pos) {
     os << "x: " << pos.x << " y: " << pos.y;