+++ src/H2DE_renderer.cpp
@@ -1,4 +1,4 @@
-#include "H2DE_renderer.h"
+#include "H2DE/H2DE_renderer.h"
 
 // INIT
 H2DE_Renderer::H2DE_Renderer(H2DE_Engine* e, std::unordered_map<std::string, SDL_Texture*>* t, std::vector<H2DE_LevelObject*>* o, std::vector<H2DE_Button*>* b) : engine(e), textures(t), objects(o), buttons(b) {
@@ -56,7 +56,9 @@ void H2DE_Renderer::render() {
 
 void H2DE_Renderer::renderObject(H2DE_LevelObject* object) {
     H2DE_LevelObjectData data = *H2DE_GetObjectData(object);
-    if (data.texture.name != "" && (*textures).find(data.texture.name) != (*textures).end()) renderObjectTexture(data);
+    data.texture->update();
+    std::string texture = data.texture->get();
+    if (texture != "" && (*textures).find(texture) != (*textures).end()) renderObjectTexture(data);
     if (data.hitboxes.size() > 0) renderObjectHitboxes(data);
 }
 
@@ -64,23 +66,25 @@ void H2DE_Renderer::renderObjectTexture(H2DE_LevelObjectData data) {
     static H2DE_Window* window = H2DE_GetWindow(engine);
     static SDL_Renderer* renderer = H2DE_GetWindowsRenderer(window);
 
-    if (data.texture.size.w == 0.0f || data.texture.size.h == 0.0f) return;
-    if (data.texture.color.a == 0) return;
+    H2DE_TextureData textureData = data.texture->getData();
+
+    if (textureData.size.w == 0.0f || textureData.size.h == 0.0f) return;
+    if (textureData.color.a == 0) return;
 
     H2DE_LevelPos posFromParents = getPosFromParents(data);
 
-    SDL_Texture* texture = (*textures)[data.texture.name];
-    SDL_Rect destRect = lvlToAbs(posFromParents, data.absolute).makeRect(lvlToAbs(data.texture.size));
+    SDL_Texture* texture = (*textures)[data.texture->get()];
+    SDL_Rect destRect = lvlToAbs(posFromParents, data.absolute).makeRect(lvlToAbs(textureData.size));
     float rotation = data.transform.rotation;
     SDL_Point pivot = lvlToAbs(data.transform.origin, data.absolute);
     SDL_RendererFlip flip = getFlip(data.transform.flip);
 
-    SDL_SetTextureColorMod(texture, data.texture.color.r, data.texture.color.g, data.texture.color.b);
-    SDL_SetTextureAlphaMod(texture, data.texture.color.a);
-    SDL_SetTextureScaleMode(texture, getScaleMode(data.texture.scaleMode));
+    SDL_SetTextureColorMod(texture, textureData.color.r, textureData.color.g, textureData.color.b);
+    SDL_SetTextureAlphaMod(texture, textureData.color.a);
+    SDL_SetTextureScaleMode(texture, getScaleMode(textureData.scaleMode));
 
-    if (data.texture.srcRect.has_value()) {
-        SDL_Rect srcRect = data.texture.srcRect.value();
+    if (textureData.srcRect.has_value()) {
+        SDL_Rect srcRect = textureData.srcRect.value();
         SDL_RenderCopyEx(renderer, texture, &srcRect, &destRect, rotation, &pivot, flip);
     } else SDL_RenderCopyEx(renderer, texture, nullptr, &destRect, rotation, &pivot, flip);
 
@@ -111,26 +115,30 @@ void H2DE_Renderer::renderObjectHitboxes(H2DE_LevelObjectData data) {
 
 void H2DE_Renderer::renderButton(H2DE_Button* button) {
     H2DE_ButtonData data = *H2DE_GetButtonData(button);
-    if (data.texture.name != "" && (*textures).find(data.texture.name) != (*textures).end()) renderButtonTexture(data);
+    data.texture->update();
+    std::string texture = data.texture->get();
+    if (texture != "" && (*textures).find(texture) != (*textures).end()) renderButtonTexture(data);
 }
 
 void H2DE_Renderer::renderButtonTexture(H2DE_ButtonData data) {
     static H2DE_Window* window = H2DE_GetWindow(engine);
     static SDL_Renderer* renderer = H2DE_GetWindowsRenderer(window);
 
-    if (data.texture.size.w == 0.0f || data.texture.size.h == 0.0f) return;
-    if (data.texture.color.a == 0) return;
+    H2DE_TextureData textureData = data.texture->getData();
+
+    if (textureData.size.w == 0.0f || textureData.size.h == 0.0f) return;
+    if (textureData.color.a == 0) return;
     
-    SDL_Texture* texture = (*textures)[data.texture.name];
-    SDL_Rect destRect = lvlToAbs(data.pos, true).makeRect(lvlToAbs(data.texture.size));
+    SDL_Texture* texture = (*textures)[data.texture->get()];
+    SDL_Rect destRect = lvlToAbs(data.pos, true).makeRect(lvlToAbs(textureData.size));
     SDL_Point pivot = { 0, 0 };
 
-    SDL_SetTextureColorMod(texture, data.texture.color.r, data.texture.color.g, data.texture.color.b);
-    SDL_SetTextureAlphaMod(texture, data.texture.color.a);
-    SDL_SetTextureScaleMode(texture, getScaleMode(data.texture.scaleMode));
+    SDL_SetTextureColorMod(texture, textureData.color.r, textureData.color.g, textureData.color.b);
+    SDL_SetTextureAlphaMod(texture, textureData.color.a);
+    SDL_SetTextureScaleMode(texture, getScaleMode(textureData.scaleMode));
 
-    if (data.texture.srcRect.has_value()) {
-        SDL_Rect srcRect = data.texture.srcRect.value();
+    if (textureData.srcRect.has_value()) {
+        SDL_Rect srcRect = textureData.srcRect.value();
         SDL_RenderCopyEx(renderer, texture, &srcRect, &destRect, 0.0f, &pivot, SDL_FLIP_NONE);
     } else SDL_RenderCopyEx(renderer, texture, nullptr, &destRect, 0.0f, &pivot, SDL_FLIP_NONE);