+++ src/H2DE_renderer_converter.cpp
@@ -1,4 +1,4 @@
-#include "H2DE_renderer.h"
+#include "H2DE/H2DE_renderer.h"
 
 H2DE_AbsPos H2DE_Renderer::lvlToAbs(H2DE_LevelPos pos, bool absolute) const {
     static H2DE_Camera* camera = H2DE_GetCamera(engine);