+++ src/H2DE_window.cpp
@@ -1,4 +1,4 @@
-#include "H2DE_window.h"
+#include "H2DE/H2DE_window.h"
 
 // INIT
 H2DE_Window::H2DE_Window(H2DE_Engine* e, H2DE_WindowData d) : engine(e), data(d) {